ef configurations init

+++ Bussiness/Abstract/IProductService.cs
@@ -11,7 +11,7 @@ namespace Bussiness.Abstract
 
         IDataResult <List<Product>> GetAllByCategoryId(int id);
 
-		IDataResult <List<Product>> GetByUnitPrice(decimal min, decimal max);
+		IDataResult <List<Product>> GetByUnitPrice(double min, double max);
 
         IDataResult <List<ProductDetailDto>> GetProductDetails();
